requirements.txt
opencv-python-headless
dlib
face_recognition
onnxruntime-gpu
numpy
scipy